+++ src/hashtbl.h
@@ -26,16 +26,16 @@ typedef struct hashtbl {
 
 struct hashnode_s {
     char *key;
-    double data;
+    size_t data;
     struct hashnode_s *next;
 };
 
 HASHTBL *hashtbl_create(hash_size size, hash_size (*hashfunc)(const char *));
 void hashtbl_destroy(HASHTBL *hashtbl);
-int hashtbl_insert(HASHTBL *hashtbl, const char *key, double data);
+int hashtbl_insert(HASHTBL *hashtbl, const char *key, size_t data);
 int hashtbl_remove(HASHTBL *hashtbl, const char *key);
-int hashtbl_update(HASHTBL *hashtbl, const char *key, double data);
-double hashtbl_get(HASHTBL *hashtbl, const char *key);
+int hashtbl_update(HASHTBL *hashtbl, const char *key, size_t data);
+size_t hashtbl_get(HASHTBL *hashtbl, const char *key);
 int hashtbl_resize(HASHTBL *hashtbl, hash_size size);
 
 #endif
